Theni Voice Using VITS/requirements.txt
TTS==0.22.0
torch
torchaudio
numpy
soundfile
librosa
ffmpeg-python
pyyaml
gradio
pandas
